Move Fuser back to proc/Fuse
